--- src/components/Authentication.jsx
@@ -1,10 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import firebaseConfig from "./firebaseConfig.jsx";
-import firebase from 'firebase/app';
-
-
-const Authentication = (props) => {
-
-}
-
-export default Authentication;
